python-agents/paymentagent/requirements.txt
langchain-openai>=0.1.0
langchain-core>=0.2.0
langchain>=0.3.0
python-dotenv>=1.0.0